SoloPhysicsHandler Implimentation

+++ src/com/print_stack_trace/voogasalad/model/engine/authoring/AbstractGameAuthorEngine.java
@@ -25,7 +25,7 @@ public interface AbstractGameAuthorEngine {
 	public void setCameraType(CameraType c) throws ElementLockedException;
 	public void setLevelCharacteristics(LevelCharacteristics levelSpecs) throws ElementLockedException;
 	public void setProgramPhysicsEngine(ProgramPhysicEngine engineType) throws ElementLockedException;
-	public void setPhysicsEngineUsingParams(int gravity, int drag, int intensity) throws ElementLockedException;
+	public void setPhysicsEngineUsingParams(float gravity, float drag, float intensity) throws ElementLockedException;
 	public void setResultOfCollision(CollisionResult result, SpriteType s1, SpriteType s2) throws ElementLockedException;
 	public void setCustomParamForCollisionType(CollisionResult result, UserDefinedCollisionParams paramType, int param) throws ElementLockedException;
 }
